requirements.txt
openai
fastapi
uvicorn
requests
python-dotenv
#google-generativeai